fix/refactor: created custom useAuth hook to prevent redirect to login page on refresh
